bot/services/court.py
"""Court cases parser for sudrf.ru."""
import logging
from typing import List, Dict, Any, Optional
import requests
from bs4 import BeautifulSoup

logger = logging.getLogger(__name__)


class CourtCasesService:
    """Service for parsing court cases from sudrf.ru."""
    
    def __init__(self):
        """Initialize court cases service."""
        self.base_url = "https://sudrf.ru"
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        })
    
    def search_cases(self, inn: str = None, company_name: str = None, page: int = 1, 
                     per_page: int = 10) -> Dict[str, Any]:
        """
        Search for court cases (best-effort parsing).
        
        Note: This is a simplified implementation as sudrf.ru has complex
        structure and may require authentication/captcha solving.
        """
        try:
            # Mock implementation - in production, this would need:
            # 1. Proper API integration or selenium for complex parsing
            # 2. Captcha solving
            # 3. Session management
            # 4. Detailed HTML parsing
            
            result = {
                'total': 0,
                'page': page,
                'per_page': per_page,
                'cases': [],
                'note': 'Для получения актуальных данных о судебных делах требуется доступ к API суда или специализированным сервисам. '
                       'Рекомендуется использовать платные API (например, Casebook.ru или аналоги).'
            }
            
            # Try to get data (best-effort)
            # This is a placeholder for actual implementation
            result['cases'] = self._parse_mock_cases(inn, company_name)
            result['total'] = len(result['cases'])
            
            return result
            
        except Exception as e:
            logger.error(f"Error searching court cases: {e}")
            return {
                'total': 0,
                'page': page,
                'per_page': per_page,
                'cases': [],
                'error': str(e),
                'note': 'Произошла ошибка при получении данных с sudrf.ru'
            }
    
    def _parse_mock_cases(self, inn: str = None, company_name: str = None) -> List[Dict]:
        """
        Mock parser for demonstration purposes.
        In production, implement actual parsing logic or API integration.
        """
        # Return empty list as we need actual API or complex parsing
        return []
    
    def get_case_details(self, case_id: str) -> Optional[Dict[str, Any]]:
        """Get detailed information about a specific case."""
        try:
            # Placeholder for actual implementation
            result = {
                'case_id': case_id,
                'note': 'Детальная информация требует интеграции с API суда'
            }
            return result
        except Exception as e:
            logger.error(f"Error getting case details: {e}")
            return None


# Global service instance
court_service = CourtCasesService()
