requirements.txt
python-telegram-bot==20.7
openai==1.12.0
reportlab==4.0.9
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp==3.9.3
python-dateutil==2.8.2
pydantic==2.6.1
flask==3.0.0
